+++ src/models/food.rs
@@ -3,5 +3,5 @@ use crate::models::coord::Coord;
 #[derive(Debug, Clone)]
 pub struct Food {
     pub pos: Coord,
-    pub restore: u8,
+    pub restore: i16,
 }
